+++ src/lib/index.ts
@@ -1,3 +1,6 @@
+export * from "./actions/component";
+export * from "./actions/stylesheet";
+
 export * from "./stores/javascript";
 export * from "./stores/pipeline";
 export * from "./stores/svelte";